更新用户、角色模块

+++ vision-parent/vision-manage/src/main/java/com/vision/service/sys/imp/SysOrganizationServiceImpl.java
@@ -8,9 +8,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.vision.exception.ServiceException;
 import com.vision.mapper.sys.SysOrganizationMapper;
+import com.vision.mapper.sys.SysUserMapper;
 import com.vision.pojo.sys.SysOrganization;
+import com.vision.pojo.sys.SysUser;
 import com.vision.service.sys.SysOrganizationService;
 import com.vision.vo.Node;
 import com.vision.vo.Node2;
@@ -19,7 +22,8 @@ import com.vision.vo.Node2;
 public class SysOrganizationServiceImpl implements SysOrganizationService{
 	@Autowired
 	private SysOrganizationMapper sysOrganizationMapper;
-
+	@Autowired
+	private SysUserMapper sysUserMapper;
 	@Override
 	public int updateOrganization(SysOrganization sysOrganization) {
 		//1.合法验证
@@ -48,17 +52,22 @@ public class SysOrganizationServiceImpl implements SysOrganizationService{
 				//2.执行删除操作
 				//2.1判定此id对应的菜单是否有子元素
 				int childCount=sysOrganizationMapper.getChildCount(organizationId);
+				
 				if(childCount>0) {
 					return -1;
-					
-					
 				}
-				
 				//2.2判定此部门是否有用户
 				//int userCount=sysUserDao.getUserCountByDeptId(id);
 				//if(userCount>0)
 				//throw new ServiceException("此部门有员工，不允许对部门进行删除");
 				//2.2判定此部门是否已经被用户使用,假如有则拒绝删除
+				QueryWrapper<SysUser> queryWrapper = new QueryWrapper<SysUser>();
+				queryWrapper.eq("organization_id", organizationId);
+				Integer selectCount = sysUserMapper.selectCount(queryWrapper);
+				if(selectCount>0) {
+					return -2;
+				}
+				
 				//2.3执行删除操作
 				int rows=sysOrganizationMapper.deleteById(organizationId);
 				if(rows==0)
@@ -73,7 +82,7 @@ public class SysOrganizationServiceImpl implements SysOrganizationService{
 				if(sysOrganization==null)
 				throw new ServiceException("保存对象不能为空");
 				if(StringUtils.isEmpty(sysOrganization.getOrganizationName()))
-				throw new ServiceException("组织不能为空");
+				throw new ServiceException("组织名不能为空");
 				//2.保存数据
 				int rows=sysOrganizationMapper.insert(sysOrganization);
 				//if(rows==1)
@@ -97,6 +106,8 @@ public class SysOrganizationServiceImpl implements SysOrganizationService{
 		
 		return result;
 	}
+	
+	
 	/**
 	 * 基于id获取组织信息
 	 * @param id
@@ -107,6 +118,5 @@ public class SysOrganizationServiceImpl implements SysOrganizationService{
 		SysOrganization organization = sysOrganizationMapper.selectById(id);
 		return organization;
 	}
-	
 	
 }
